Lift repo-sync-test types up into a types/ directory
